utils/dashboard: Split dashboard tab functionality from module
