+++ lib/css/modules/_dashboard.scss
@@ -1,8 +1,3 @@
-// this first line hides the "you need RES 4.0+ to view the dashboard" link
-.id-t3_qi5iy {
-	display: none;
-}
-
 .RESDashboardComponent a.widgetPath,
 .addNewWidget,
 .editWidget {
@@ -295,86 +290,5 @@ ul.widgetStateButtons {
 		.res-icon {
 			margin: 3px;
 		}
-	}
-}
-
-#userTaggerContents .show {
-	display: inline-block;
-}
-
-#tagPageControls {
-	display: inline-block;
-	margin: 0 0 0 10px;
-}
-
-// some styles to override /r/dashboard stylesheet.
-#userTaggerContents,
-#newCommentsContents {
-	margin-right: 305px;
-	font-size: small;
-}
-
-#userTaggerContents table,
-#newCommentsContents table {
-	margin: 0;
-	width: 100%;
-}
-
-#userTaggerTable th,
-#newCommentsTable th,
-#userTaggerTable td,
-#newCommentsTable td {
-	min-width: 0;
-	padding: 5px 8px;
-	border: 1px solid rgb(200, 200, 200);
-}
-
-#newCommentsTable th,
-#userTaggerTable th,
-#newCommentsTable td {
-	white-space: nowrap;
-}
-
-#newCommentsTable td:first-child,
-#newCommentsTable th:first-child {
-	width: auto;
-	white-space: normal;
-}
-
-#newCommentsTable .deleteIcon,
-#userTaggerTable .deleteIcon {
-	color: rgba(0, 0, 0, 0);
-}
-
-#newCommentsTable tr:hover .deleteIcon,
-#userTaggerTable tr:hover .deleteIcon {
-	color: #369;
-}
-
-body.res-dashboard {
-	.id-t3_8qcn09 {
-		display: none;
-	}
-
-	.tabmenu li:not(.res-tabmenu-button),
-	.pagename {
-		display: none;
-	}
-
-	.side {
-		display: none;
-	}
-
-	.res-dashboard-tab {
-		font-variant: small-caps;
-		font-size: 1.2em;
-		vertical-align: bottom;
-
-		// We want the tabs to act as a radio, i.e. don't allow deselect the currently selected
-		&.selected a:first-child { pointer-events: none; }
-	}
-
-	.tabmenu {
-		display: inline-block !important;
 	}
 }
