+++ lib/css/modules/_userTagger.scss
@@ -137,3 +137,12 @@ a.voteWeight {
 		color: #333;
 	}
 }
+
+#userTaggerContents .show {
+	display: inline-block;
+}
+
+#tagPageControls {
+	display: inline-block;
+	margin: 0 0 0 10px;
+}
