+++ lib/modules/dashboard.js
@@ -7,7 +7,7 @@ import { Module } from '../core/module';
 import {
 	Alert,
 	CreateElement,
-	BodyClasses,
+	addDashboardTab,
 	currentSubreddit,
 	formatDateTime,
 	isCurrentSubreddit,
@@ -19,7 +19,6 @@ import {
 import { i18n, Storage, ajax } from '../environment';
 import * as Menu from './menu';
 import * as Notifications from './notifications';
-import * as SettingsNavigation from './settingsNavigation';
 
 export const module: Module<*> = new Module('dashboard');
 
@@ -111,50 +110,13 @@ module.beforeLoad = () => {
 };
 
 module.contentStart = () => {
-	addTab('dashboardContents', 'My dashboard', module.moduleID, addDashboardTab);
+	addDashboardTab('dashboardContents', 'My dashboard', module.moduleID, addDashboardFunctionality);
 
 	const subreddit = currentSubreddit();
 	if (module.options.dashboardShortcut.value && subreddit) {
 		addDashboardShortcuts(subreddit);
 	}
 };
-
-const initialTabID = location.hash.replace('#', '') || 'dashboardContents';
-let selectedTabMenuItem;
-
-export function addTab(tabID: string, tabName: string, tabModuleId?: string, callback: JQuery => any) {
-	if (!isCurrentSubreddit('dashboard')) return;
-
-	BodyClasses.add('res-dashboard');
-
-	let $tabPage;
-
-	const tabMenuItem = CreateElement.tabMenuItem({
-		text: tabName,
-		className: 'res-dashboard-tab',
-		order: -1,
-		onChange: active => {
-			if (active) {
-				if (selectedTabMenuItem) selectedTabMenuItem.click();
-				selectedTabMenuItem = tabMenuItem;
-				$tabPage = $('<div class="dashboardPane">').appendTo('#siteTable.linklisting');
-				callback($tabPage);
-				location.hash = tabID;
-			} else {
-				$tabPage.remove();
-			}
-		},
-	});
-
-	// if a module id has been provided then we automatically append the gear link
-	if (tabModuleId) {
-		const gear = string.html`${string.safe(SettingsNavigation.makeUrlHashLink(
-			tabModuleId, undefined, ' ', 'gearIcon'))}`;
-		tabMenuItem.after(gear);
-	}
-
-	if (tabID === initialTabID) tabMenuItem.click();
-}
 
 const MAX_ROWS = 100;
 
@@ -177,9 +139,9 @@ async function getLatestWidgets() {
 		.filter(Boolean);
 }
 
-async function addDashboardTab($tabPage) {
+async function addDashboardFunctionality(tabPage) {
 	await initialWidgetLoad();
-	attachAddComponent($tabPage);
+	attachAddComponent($(tabPage));
 	attachEditComponent();
 	initUpdateQueue();
 }
