+++ lib/modules/newCommentCount.js
@@ -8,6 +8,7 @@ import {
 	MINUTE,
 	Alert,
 	Thing,
+	addDashboardTab,
 	escapeHTML,
 	execRegexes,
 	filterMap,
@@ -24,7 +25,6 @@ import {
 	waitForEvent,
 } from '../utils';
 import { Storage, isPrivateBrowsing } from '../environment';
-import * as Dashboard from './dashboard';
 import * as Notifications from './notifications';
 
 export const module: Module<*> = new Module('newCommentCount');
@@ -118,7 +118,7 @@ module.contentStart = () => {
 		$('.commentarea .panestack-title').append(subscriptionButton);
 	}
 
-	Dashboard.addTab('newCommentsContents', 'My Subscriptions', module.moduleID, addDashboardFunctionality);
+	addDashboardTab('newCommentsContents', 'My Subscriptions', module.moduleID, addDashboardFunctionality);
 
 	maybePruneOldEntries('newCommentCount', entryStorage, parseInt(module.options.cleanComments.value, 10));
 };
@@ -351,7 +351,8 @@ function createButton(id, type) {
 	return { button: $button.get(0), promise: waitForEvent($button.get(0), 'click').then(action) };
 }
 
-function addDashboardFunctionality($tabPage) {
+function addDashboardFunctionality(tabPage) {
+	const $tabPage = $(tabPage);
 	const $openOnReddit = $('<a href="#" id="openOnReddit">as reddit link listing</a>');
 	$openOnReddit.click(event => {
 		event.preventDefault();
