+++ lib/modules/saveComments.js
@@ -8,6 +8,7 @@ import * as Modules from '../core/modules';
 import {
 	CreateElement,
 	WEEK,
+	addDashboardTab,
 	formatDate,
 	formatDateDiff,
 	formatDateTime,
@@ -19,7 +20,6 @@ import {
 	Thing,
 } from '../utils';
 import { Storage, i18n } from '../environment';
-import * as Dashboard from './dashboard';
 import * as KeyboardNav from './keyboardNav';
 import * as Notifications from './notifications';
 import * as SettingsNavigation from './settingsNavigation';
@@ -56,7 +56,7 @@ module.contentStart = () => {
 		});
 	}
 
-	Dashboard.addTab('savedComments', 'Saved Comments', module.moduleID, async $tabPage => $tabPage.append(await drawSavedComments()));
+	addDashboardTab('savedComments', 'Saved Comments', module.moduleID, async tabPage => $(tabPage).append(await drawSavedComments()));
 };
 
 const saveElement = (e => () => preventCloning(e().cloneNode(true)))(_.once(() => {
