+++ lib/modules/userTagger.js
@@ -7,8 +7,9 @@ import * as Options from '../core/options';
 import {
 	Alert,
 	CreateElement,
-	Thing,
 	SelectedThing,
+	Thing,
+	addDashboardTab,
 	downcast,
 	isPageType,
 	usernameSelector,
@@ -169,7 +170,7 @@ module.contentStart = () => {
 
 	registerCommandLine();
 
-	Dashboard.addTab('userTaggerContents', i18n('userTaggerMyUserTags'), module.moduleID, addDashboardFunctionality);
+	addDashboardTab('userTaggerContents', i18n('userTaggerMyUserTags'), module.moduleID, addDashboardFunctionality);
 };
 
 function applyFromElement(element) {
@@ -667,8 +668,8 @@ const filters = [
 	{ get name() { return i18n('userTaggerTaggedUsers'); }, filter: v => v.text },
 ];
 
-async function addDashboardFunctionality($tabPage) {
-	// populate the contents of the tab
+async function addDashboardFunctionality(tabPage) {
+	const $tabPage = $(tabPage);
 	const $showDiv = $(string.html`<div class="show">${i18n('userTaggerShow')} </div>`)
 		.appendTo($tabPage);
