+++ lib/utils/index.js
@@ -44,6 +44,10 @@ export {
 } from './color';
 
 export {
+	addDashboardTab,
+} from './dashboard';
+
+export {
 	addCSS,
 	click,
 	elementInViewport,
